Move remaining component files into separate directories
