+++ static/js/components/FoldableRow/index.jsx
@@ -9,7 +9,7 @@
  * */
 
 import React, { Component, PropTypes } from 'react';
-import colorScheme from './colorscheme';
+import colorScheme from '../colorscheme';
 
 const cs = colorScheme;
